django/requirements.txt
Django>=3.0,<4.1
psycopg2-binary
djangorestframework
pillow
drf_yasg
djangorestframework-simplejwt
django-cors-headers
django-filter
# facebook-sdk
# Faker
# future
# google-api-core
# google-api-python-client
# google-auth
# google-auth-httplib2
# googleapis-common-protos
# gunicorn
# httplib2